requirements.txt
openai
streamlit
pillow
ipykernel
jupyter
python-dotenv
pandas
